refactor organizations api to remove bb_ code BB-9402 #resolve
