--- bluebottle/bb_organizations/permissions.py
@@ -1,16 +1,0 @@
-from .models import BaseOrganization
-
-from rest_framework import permissions
-
-
-class IsOrganizationMember(permissions.BasePermission):
-    """
-    Allows access only to organization members
-    """
-
-    def has_object_permission(self, request, view, obj):
-        if isinstance(obj, BaseOrganization):
-            return request.user.id in obj.members.values_list('user_id',
-                                                              flat=True)
-        return request.user.id in obj.organization.members.values_list(
-            'user_id', flat=True)
