--- bluebottle/bb_organizations/urls/api.py
@@ -1,15 +1,0 @@
-from django.conf.urls import url
-
-from ..views import (
-    ManageOrganizationList, ManageOrganizationDetail, OrganizationDetail,
-    OrganizationList)
-
-urlpatterns = [
-    url(r'^$', OrganizationList.as_view(), name='organization_list'),
-    url(r'^(?P<pk>\d+)$', OrganizationDetail.as_view(),
-        name='organization_detail'),
-    url(r'^manage/$', ManageOrganizationList.as_view(),
-        name='manage_organization_list'),
-    url(r'^manage/(?P<pk>\d+)$', ManageOrganizationDetail.as_view(),
-        name='manage_organization_detail'),
-]
