--- bluebottle/bb_organizations/urls/main.py
@@ -1,9 +1,0 @@
-from django.conf.urls import url
-
-from ..views import RegistrationDocumentDownloadView
-
-urlpatterns = [
-    url(r'^organizations/(?P<pk>\d+)/$',
-        RegistrationDocumentDownloadView.as_view(),
-        name='organization_registration_download'),
-]
