+++ bluebottle/organizations/serializers.py
@@ -1,19 +1,14 @@
 from rest_framework import serializers
 
-from bluebottle.organizations.models import Organization
 from bluebottle.utils.serializers import URLField
-from bluebottle.organizations.models import OrganizationContact
-
-
-class OrganizationContactSerializer(serializers.ModelSerializer):
-    name = serializers.CharField(required=True)
-
-    class Meta:
-        model = OrganizationContact
-        fields = ('name', 'email', 'phone')
+from bluebottle.organizations.models import Organization, OrganizationContact
 
 
 class OrganizationSerializer(serializers.ModelSerializer):
+    slug = serializers.SlugField(required=False, allow_null=True)
+    name = serializers.CharField(required=True)
+    website = URLField(required=False, allow_blank=True)
+    email = serializers.EmailField(required=False, allow_blank=True)
     contacts = serializers.SerializerMethodField()
 
     def get_contacts(self, obj):
@@ -30,22 +25,11 @@ class OrganizationSerializer(serializers.ModelSerializer):
         model = Organization
         fields = ('id', 'name', 'slug', 'address_line1', 'address_line2',
                   'city', 'state', 'country', 'postal_code', 'phone_number',
-                  'website', 'email', 'contacts')
+                  'website', 'email', 'contacts', 'partner_organizations',
+                  'created', 'updated')
 
 
-class ManageOrganizationSerializer(OrganizationSerializer):
-    slug = serializers.SlugField(required=False, allow_null=True)
-    name = serializers.CharField(required=True)
-    website = URLField(required=False, allow_blank=True)
-    email = serializers.EmailField(required=False, allow_blank=True)
-
-    class Meta:
-        model = Organization
-        fields = OrganizationSerializer.Meta.fields + ('partner_organizations',
-                                                       'created', 'updated')
-
-
-class ManageOrganizationContactSerializer(serializers.ModelSerializer):
+class OrganizationContactSerializer(serializers.ModelSerializer):
     name = serializers.CharField(required=True)
     email = serializers.EmailField(required=True)
     phone = serializers.CharField(required=False, allow_blank=True)
